feat: Speed up performance due to using dynamic imports with codesplitting
