+++ packages/midi-bricks-electron/app/electron.js
@@ -167,7 +167,7 @@ async function createWindow() {
 
   const url = isDev
     ? 'http://localhost:3000/'
-    : `file://${path.join(__dirname, '../build/index.html')}`
+    : `file://${path.join(__dirname, '../index.html')}`
 
   win.loadURL(url)
   isAllowedToUpdate && sendStatusToWindow('Software-Updates enabled.')
