+++ packages/midi-bricks/src/App.jsx
@@ -23,7 +23,7 @@ function App() {
           anchor={'left'}
           open={isMobileOpen}
           classes={{
-            paper: classes.drawerPaper,
+            paper: classes.drawerPaper
           }}
           onClose={() => setIsMobileOpen(!isMobileOpen)}
         >
@@ -48,8 +48,7 @@ function styles(theme) {
     root: {
       width: '100%',
       height: '100%',
-      zIndex: 1,
-
+      zIndex: 1
     },
     appFrame: {
       position: 'relative',
@@ -79,8 +78,7 @@ function styles(theme) {
     content: {
       backgroundColor: theme.palette.background.default,
       width: '100%',
-      marginTop: theme.spacing(1),
-
+      marginTop: theme.spacing(1)
     }
   }
 }
