+++ packages/midi-bricks/src/pages/Home.jsx
@@ -1,6 +1,5 @@
 import { makeStyles, useTheme } from '@material-ui/styles'
-
-import React, { useEffect } from 'react'
+import React, { useEffect, Suspense } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
@@ -8,15 +7,10 @@ import { initApp } from '../actions/init.js'
 import { Actions as MidiSliderActions } from '../actions/slider-list.js'
 import { Actions as ViewStuff } from '../actions/view-settings.js'
 import ChannelStripList from '../components/channel-strip-list/ChannelStripList'
-import GlobalSettingsPage from './GlobalSettingsPage.jsx'
-import MidiDriversSettingsPage from './MidiDriversSettingsPage'
 import ApplicationSettingsPage from '../components/ApplicationSettings'
 import { PAGE_TYPES } from '../reducers'
 
-export default connect(
-  mapStateToProperties,
-  mapDispatchToProperties
-)(Home)
+export default connect(mapStateToProperties, mapDispatchToProperties)(Home)
 
 Home.propTypes = {
   classes: PropTypes.object,
@@ -26,7 +20,7 @@ Home.propTypes = {
 
 function Home(props) {
   const theme = useTheme()
-  const classes = makeStyles(styles.bind(this,theme), { useTheme: true })()
+  const classes = makeStyles(styles.bind(this, theme), { useTheme: true })()
 
   const {
     viewSettings: { isLiveMode = false, pageType = PAGE_TYPES.HOME_MODE },
@@ -46,18 +40,32 @@ function Home(props) {
 
   const preventScrollStyle = isLiveMode
     ? {
-      height: 'calc(100vh - 66px)',
-      overflowY: 'hidden'
-    }
+        height: 'calc(100vh - 66px)',
+        overflowY: 'hidden'
+      }
     : {
-      height: 'calc(100vh - 66px - 64px)',
-      overflowY: 'hidden'
-    }
+        height: 'calc(100vh - 66px - 64px)',
+        overflowY: 'hidden'
+      }
 
   if (pageType === PAGE_TYPES.GLOBAL_MODE) {
-    return <GlobalSettingsPage />
+    const GlobalSettingsPage = React.lazy(() =>
+      import('./GlobalSettingsPage.jsx')
+    )
+    return (
+      <Suspense fallback={<div>Loading...</div>}>
+        <GlobalSettingsPage />
+      </Suspense>
+    )
   } else if (pageType === PAGE_TYPES.MIDI_DRIVER_MODE) {
-    return <MidiDriversSettingsPage />
+    const MidiDriversSettingsPage = React.lazy(() =>
+      import('./MidiDriversSettingsPage.jsx')
+    )
+    return (
+      <Suspense fallback={<div>Loading...</div>}>
+        <MidiDriversSettingsPage />
+      </Suspense>
+    )
   } else if (pageType === PAGE_TYPES.VIEW_SETTINGS_MODE) {
     return <ApplicationSettingsPage />
   } else if (pageType === PAGE_TYPES.HOME_MODE) {
@@ -72,10 +80,7 @@ function Home(props) {
   }
 }
 
-function mapStateToProperties({
-  viewSettings,
-  sliders: { isMidiFailed }
-}) {
+function mapStateToProperties({ viewSettings, sliders: { isMidiFailed } }) {
   return {
     viewSettings,
     isMidiFailed
@@ -104,7 +109,7 @@ function styles(theme) {
     },
     noMidiTypography: {
       textAlign: 'center',
-      paddingTop: theme.spacing(4),
+      paddingTop: theme.spacing(4)
     }
   }
 }
